+++ documenteer/conf/pipelines.py
@@ -123,6 +123,7 @@ extensions = [
     "documenteer.sphinxext",
     "documenteer.sphinxext.lssttasks",
     "documenteer.ext.autocppapi",
+    "documenteer.ext.autodocreset",
     "sphinx_click",
 ]
 
@@ -298,9 +299,6 @@ html_file_suffix = ".html"
 
 # Language to be used for generating the HTML full-text search index.
 html_search_language = "en"
-
-html_extra_path = [""]
-
 
 # ============================================================================
 # #AUTOMODAPI automodapi and autodoc configuration